+++ libraries/RunningAverage/CHANGELOG.md
@@ -6,10 +6,23 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.4.6] - 2024-06-15
+- Fix #30, add **float getSum()** (thanks to heidnerd)
+- Fix #31, add **float getStandardDeviationLast(uint16_t count)** (thanks to alvaro-oliver)
+- added code to detect inner array == NULL in more functions (return NAN).
+  - changed return type **bool clear()**
+  - changed return type **bool addValue(..)**
+  - changed return type **bool fillValue(..)**
+  - changed return type **bool setPartial(..)**
+- moved performance.txt to performance sketch
+- update unit test
+- update keywords.txt
+- update readme.md
+- minor edits.
+
 ## [0.4.5] - 2024-01-05
 - fix URL in examples
 - minor edits
-
 
 ## [0.4.4] - 2023-10-18
 - update readme.md badges
@@ -29,7 +42,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - minor edits
 
 ## [0.4.1] - 2021-11-22
-- updated buil-CI, readme, badges
+- updated build-CI, readme, badges
 - add getAverageLast() functions.
 
 ## [0.4.0] - 2021-05-18
@@ -56,7 +69,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.2.15] - 2020-01-17
 - fix overflow in getValue - see issue #139
-
 
 ## [0.2.14] - 2020-01-15
 - added getValue(n) to retrieve elements in order of addition - see issue #132
@@ -107,7 +119,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.2.01] - 2012-11-21
 - refactored
 
-## [0.2.00] - 2012-??-??  
+## [0.2.00] - 2012-??-??
 - Yuval Naveh added trimValue (found on web)
 - http://stromputer.googlecode.com/svn-history/r74/trunk/Arduino/Libraries/RunningAverage/RunningAverage.cpp